Removed Task.Run() from WorkoutService -> bad practice. Refactored workout entity -> needs a migration for db

+++ ExerciseWebsite/Entities/Workout.cs
@@ -12,7 +12,7 @@ namespace ExerciseWebsite.Entities
         public string Title { get; set; }
         public string Description { get; set; }
         public double AvgDifficulty { get; set; }
-        public int Rating { get; set; }
+        public double Rating { get; set; } // must start over to alter column in Db sadly
         public int RatingCount { get; set; }
         public DateTime DateCreated { get; set; }
     }
